Renamed files and fixed unified selection tree
