+++ src/frontend-samples/tree-samples/presentation-tree/PresentationTree.tsx
@@ -7,7 +7,7 @@ import { IModelConnection } from "@bentley/imodeljs-frontend";
 import { ControlledTree, SelectionMode, useTreeEventsHandler, useVisibleTreeNodes } from "@bentley/ui-components";
 import { usePresentationTreeNodeLoader } from "@bentley/presentation-components";
 import { Ruleset } from "@bentley/presentation-common";
-import { ReloadableConnection } from "../../../Components/ReloadableComponent/ReloadableConnection";
+import { ReloadableConnection } from "../../../Components/GenericReloadableComponent/GenericReloadableComponent";
 
 const PAGING_SIZE = 20;
 const RULESET_TREE_HIERARCHY: Ruleset = require("../TreeHierarchy.json"); // tslint:disable-line: no-var-requires
@@ -16,7 +16,7 @@ export interface PresentationTreeProps {
   imodel: IModelConnection;
 }
 
-export class PresentationTreeSample extends React.Component<{ iModelName: string }, { iModel: IModelConnection }> {
+export class PresentationTreeSample extends React.Component<{ iModelName: string, iModelSelector: React.ReactNode }, { iModel: IModelConnection }> {
 
   public getControlPane() {
     return (
@@ -25,13 +25,14 @@ export class PresentationTreeSample extends React.Component<{ iModelName: string
           <div className="sample-instructions">
             <span>Data in this tree is loaded using Presentation rules.</span>
           </div>
+          {this.props.iModelSelector}
         </div>
       </>
     );
   }
 
-  public static async setup(iModelName: string) {
-    return <PresentationTreeSample iModelName={iModelName}></PresentationTreeSample>;
+  public static async setup(iModelName: string, iModelSelector: React.ReactNode) {
+    return <PresentationTreeSample iModelName={iModelName} iModelSelector={iModelSelector}></PresentationTreeSample>;
   }
 
   public onIModelReady = (imodel: IModelConnection) => {
